auth refactor
